Added docker log symlinks to localhost /var/log/*
Our container application logs are not being stored in their respective default locations. Instead they are being dynamically generated outside of the container. This log can be viewed by using "docker logs <container_name>", but to help remove this minor layer of abstraction, I've created symlinks from these files to /var/log/service on the localhost when the docker container is created.

I have also created volume bind mounts that map to the same /var/log/service locations for each service that does not currently have logging enabled, so that if "in-container" logs are to be generated, they can be easily shared with the localhost.

For containers that are configured to store logs in specific locations (/usr/share/service/logs), I have mapped those areas to /var/log/service/ on the localhost. Elasticsearch and Filebeats are currently set up in this manner.